refactor: remove dead comments and an internal barrel file
- Drop commented-out IMAGE_EXCLUDED_PROPS block in widgetPropFilter.
- Drop commented-out code in useSelectionOperations and LGraphCanvas.
- Remove src/components/ui/toggle-group/index.ts barrel; import the two
  components directly per the no-internal-barrels rule.

Left src/types/nodeSource.ts: it re-exports an external package as a stable
internal alias for 15 call sites, not an internal barrel.
