--- src/components/ui/toggle-group/index.ts
@@ -1,2 +1,0 @@
-export { default as ToggleGroup } from './ToggleGroup.vue'
-export { default as ToggleGroupItem } from './ToggleGroupItem.vue'
